feat(inc_search): remove default reverse for IncSearch hightlight group
Vim default IncSearch group defines gui="reverse" in order to have one group for the search and
one group for the incremental search, the latter will have the colors swapped to be recognizable.

themer.lua should remove such default as in the configuration it clearly says `fg` and `bg` for the `inc_search` group,
so one to have the correct visualization has to swap the foreground and background.

BREAKING_CHANGE(inc_search): swap `fg` and `bg` in `inc_search` group