feat(mapper): add config.diagnostic_underline
Mostly personal preference, but I prefer a straight line instead of the curly one for the diagnostic,
I wanted to change it to do so but I saw there was no way to do so.

This PR adds `diagnostic_underline` to the config table so it'll work accross themes.

This should not break any configuration as I've put the old value as default.